requirements.txt

torch
torchvision
mlflow
matplotlib
numpy
pillow
